Deleted unused imports and changed id type from int to long

+++ src/main/java/com/oleksandr/remitly/swiftapi/Repository/BankNameRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BankNameRepository extends JpaRepository<BankName, Integer> {
+public interface BankNameRepository extends JpaRepository<BankName, Long> {
     Optional<BankName> findByName(String bankName);
 
     Optional<BankName> findByNameAndCountry(String bankName, Country country);
