+++ src/main/java/com/oleksandr/remitly/swiftapi/Repository/CityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CityRepository extends JpaRepository<City, Integer> {
+public interface CityRepository extends JpaRepository<City, Long> {
     Optional<City> findByName(String cityName);
 
     Optional<City> findByNameAndCountry(String cityName, Country country);
