+++ src/main/java/com/oleksandr/remitly/swiftapi/Repository/TimezoneRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TimezoneRepository extends JpaRepository<Timezone, Integer> {
+public interface TimezoneRepository extends JpaRepository<Timezone, Long> {
     Optional<Timezone> findByTimezoneName(String timezoneName);
 }
